requirements.txt
flask
flask_restful
request
Flask-CORS
apscheduler
pymysql
kafka-python
SQLAlchemy
pyyaml
scrapy
requests
bcrypt
pymongo
flask-jwt-extended
boto3
logger
stripe
pystache




